api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.29.0",
    "fastapi[standard]>=0.109.0",
    "google-generativeai>=0.8.0",
    "markdown>=3.5.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.7.1",
    "sentence-transformers>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "uvicorn[standard]>=0.27.0",
]
